ESP32: Fixed the type of cpuint variables in esp32_emac.c esp32_i2c.c esp32_spi.c esp32_spi_slave.c

+++ arch/xtensa/src/esp32/esp32_i2c.c
@@ -152,7 +152,7 @@ struct esp32_i2c_priv_s
   uint8_t msgid;               /* Current message ID */
   ssize_t bytes;               /* Processed data bytes */
 
-  uint8_t  cpuint;             /* CPU interrupt assigned to this I2C */
+  int     cpuint;              /* CPU interrupt assigned to this I2C */
 
   uint32_t error;              /* I2C transform error */
